+++ src/crypto.rs
@@ -1,0 +1,5 @@
+pub mod encrypt;
+pub mod decrypt;
+pub mod key;
+pub mod primitives;
+pub mod streams;
