+++ src/global/secret.rs
