+++ src/subcommands.rs
@@ -9,12 +9,21 @@ use crate::global::parameters::{
     parameter_handler,
 };
 
+pub mod encrypt;
+pub mod decrypt;
+pub mod erase;
+pub mod hashing;
+pub mod list;
+pub mod key;
+pub mod prompt;
+pub mod header;
+
 pub fn encrypt(sub_matches: &ArgMatches) -> Result<()> {
     let params = parameter_handler(sub_matches)?;
     let algorithm = encrypt_additional_params(sub_matches)?;
 
     // stream mode is the default - it'll redirect to memory mode if the file is too small
-    crate::encrypt::stream_mode(
+    encrypt::stream_mode(
         &get_param("input", sub_matches)?,
         &get_param("output", sub_matches)?,
         &params,
@@ -27,7 +36,7 @@ pub fn decrypt(sub_matches: &ArgMatches) -> Result<()> {
     let header = decrypt_additional_params(sub_matches)?;
 
     // stream decrypt is the default as it will redirect to memory mode if the header says so
-    crate::decrypt::stream_mode(
+    decrypt::stream_mode(
         &get_param("input", sub_matches)?,
         &get_param("output", sub_matches)?,
         &header,
@@ -38,5 +47,5 @@ pub fn decrypt(sub_matches: &ArgMatches) -> Result<()> {
 pub fn erase(sub_matches: &ArgMatches) -> Result<()> {
     let passes = erase_params(sub_matches)?;
 
-    crate::erase::secure_erase(&get_param("input", sub_matches)?, passes)
+    erase::secure_erase(&get_param("input", sub_matches)?, passes)
 }
