+++ src/subcommands/hashing.rs
